+++ src/main/java/benben/Parser.java
@@ -1,5 +1,5 @@
 
-package BenBen;
+package benben;
 public class Parser {
 
 //    private static Storage storage;
