+++ src/main/java/benben/Storage.java
@@ -1,4 +1,4 @@
-package BenBen;
+package benben;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -38,6 +38,7 @@ public class Storage {
             //System.out.println(file.exists());
             if (!file.exists()) {
                 boolean isCreated = file.createNewFile();
+                return list;
                 //System.out.println(isCreated + "new file created");
             }
             Scanner sc = new Scanner(file);
@@ -45,7 +46,7 @@ public class Storage {
             while (sc.hasNext()) {
                 boolean canRead = false;
                 String task = sc.nextLine();
-                //System.out.println(task);
+                System.out.println(task);
                 String[] strSplit = task.split("\\|");
 
                 for (int i = 0; i < strSplit.length; i++) {
@@ -57,38 +58,51 @@ public class Storage {
                     Task nextTask = new Todo(strSplit[2]);
 //                    System.out.println(strSplit[1]);
 //                    System.out.println(strSplit[1].equals("1"));
-                    if (strSplit[1].equals("1")) {
+                    if (strSplit[1].startsWith("1")) {
                         nextTask.mark();
+                        list.add(nextTask);
+                        canRead = true;
+                    } else if (strSplit[1].startsWith("0")) {
+                        list.add(nextTask);
+                        canRead = true;
                     }
-                    list.add(nextTask);
-                    canRead = true;
 
                 }
                 if (strSplit[0].startsWith("D") && strSplit.length == 4) {
                     Task nextTask = new Deadline(strSplit[2], strSplit[3]);
                     if (strSplit[1].startsWith("1")) {
                         nextTask.mark();
+                        list.add(nextTask);
+                        canRead = true;
+                    } else if (strSplit[1].startsWith("0")) {
+                        list.add(nextTask);
+                        canRead = true;
                     }
-                    list.add(nextTask);
-                    canRead = true;
                 }
                 if (strSplit[0].startsWith("E") && strSplit.length == 5) {
                     Task nextTask = new Event(strSplit[2], strSplit[3], strSplit[4]);
                     if (strSplit[1].startsWith("1")) {
                         nextTask.mark();
+                        list.add(nextTask);
+                        canRead = true;
+                    } else if (strSplit[1].startsWith("0")) {
+                        list.add(nextTask);
+                        canRead = true;
                     }
-                    list.add(nextTask);
-                    canRead = true;
+
                 }
                 if (!canRead) {
                     throw new BenBenException("The file content is corrupted, please report this to admin");
                 }
             }
+            //System.out.println(list.size());
             return list;
         } catch (FileNotFoundException e) {
             throw new BenBenException("The local file is not found in the directory");
         } catch (IOException e) {
             throw new BenBenException("IOException found!" + e.getMessage());
+        } catch (NullPointerException e) {
+            throw new BenBenException("The local file does not exits");
         }
     }
 }
