+++ src/main/java/benben/Ui.java
@@ -1,4 +1,4 @@
-package BenBen;
+package benben;
 public class Ui {
     private static final String line ="_______________________________________\n";
