+++ src/test/java/benben/StorageTest.java
@@ -1,0 +1,2 @@
+package benben;public class StorageTest {
+}
